requirements.txt
-i https://pypi.org/simple
asn1crypto==0.24.0
cached-property==1.5.1
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
click==7.0
cryptography==2.3.1
defusedxml==0.5.0
dnspython==1.15.0
exchangelib==1.12.0
flask==1.0.2
future==0.16.0
idna==2.7
isodate==0.6.0
itsdangerous==0.24
jinja2==2.10
lxml==4.2.5
markupsafe==1.0
ntlm-auth==1.2.0
pycparser==2.19
pygments==2.2.0
python-dateutil==2.7.3
python-magic==0.4.15
pytz==2018.5
requests-ntlm==1.1.0
requests==2.20.0
six==1.11.0
python-slugify==1.2.6
thehive4py==1.5.1
tzlocal==1.5.1
urllib3==1.23
werkzeug==0.14.1
PyYAML==5.1.1
redis==3.3.8